Support CSS selectors in theme files (#5)
* Install rehype-add-classes

* Redo rendering logic to use rehype-add-classes plugin

As CSS selectors are a superset of HTML tags, we can safely remove our
custom component handler in favour of simply using the plugin for
all the injections of the Tailwind classes.

* Refactor class mapping logic

* Migrate class mapping logic under theme utilities
* Add TODOs where applicable

* Support more CSS selectors in theme files

See https://github.com/martypdx/rehype-add-classes for more info. The
rehype-add-classes plugin uses a very old version of hast-util-select as
a transitive dependency, so we manually bump the version in our
package.json using selective dependency resolution in order to support
additional CSS selectors like `:has()`.